Refactor source layout into modules
