+++ src/client/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_executable(kv_client client.cpp)
+target_include_directories(kv_client PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/..)
