+++ src/map/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(map_lib INTERFACE)
+target_include_directories(map_lib INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/..)
