+++ src/utils/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(utils_lib STATIC
+    file_buffer_writer.cpp
+    spdk_buffer_writer.cpp
+)
+target_include_directories(utils_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/..)
