+++ src/utils/i_writable_buffer.h
