+++ src/utils/spdk_buffer_writer.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "i_writable_buffer.h"
+#include "utils/i_writable_buffer.h"
 
 class SPDKBufferWriter : public IWritableBuffer {
 public:
