+++ tests/unit/server_test.cpp
@@ -1,4 +1,4 @@
-#include "server_impl.h"
+#include "server/kv_server.h"
 #include "storage/memtable_factory.h"
 #include <nlohmann/json.hpp>
 #include <chrono>
