+++ tests/unit/spdk_buffer_writer_test.cpp
@@ -1,4 +1,4 @@
-#include "spdk_buffer_writer.h"
+#include "utils/spdk_buffer_writer.h"
 #include <cstdint>
 #include <cstring>
 #include <gtest/gtest.h>
